funcionalidades do usuario

+++ codigo/javascript/desaparecidos_localizados.js
@@ -1,4 +1,6 @@
-const apiUrl = "https://c75b6410-fffa-4b73-b82c-d1c21ec77f4a-00-3360t4cd2jp1v.picard.replit.dev/animais_perdidos";
+const apiUrl = "https://bc8bb33f-6175-4214-998c-292c322364a2-00-2ddr60lv3tm7s.worf.replit.dev/animais_perdidos";
+const usersApiUrl = "https://bc8bb33f-6175-4214-998c-292c322364a2-00-2ddr60lv3tm7s.worf.replit.dev/users";
+
 document.addEventListener("DOMContentLoaded", function () {
   const menuIcon = document.querySelector(".mobile-menu-icon button");
   const menu = document.querySelector(".menu");
@@ -22,6 +24,8 @@ function loadAndDisplayPets(filterStatus = null, filterTipo = null) {
           (!filterStatus || pet.status.toLowerCase() === filterStatus.toLowerCase()) &&
           (!filterTipo || pet.especie.toLowerCase() === filterTipo.toLowerCase())
       );
+
+      const userId = parseInt(localStorage.getItem('userId'));
 
       filteredPets.forEach((pet) => {
         const content = document.createElement("div");
@@ -53,75 +57,69 @@ function loadAndDisplayPets(filterStatus = null, filterTipo = null) {
         const contacts = document.createElement("strong");
         contacts.textContent = "Contatos: " + pet.contatos;
       
-        const editButton = document.createElement("button");
-        editButton.setAttribute("type", "button");
-        editButton.classList.add("btn", "btn-custom-green");
-        editButton.setAttribute("data-bs-toggle", "modal");
-        editButton.setAttribute("data-bs-target", "#editModal");
-        editButton.textContent = "Editar";
-      
-        editButton.addEventListener("click", function () {
-          document.getElementById("editId").value = pet.id;
-          document.getElementById("status").value = pet.status;
-          document.getElementById("especie").value = pet.especie;
-          document.getElementById("genero").value = pet.genero;
-          document.getElementById("nome").value = pet.nome;
-          document.getElementById("endereco").value = pet.endereco;
-          document.getElementById("contatos").value = pet.contatos;
-          document.getElementById("descricao").value = pet.descricao;
-          document.getElementById("imagemUrl").value = pet.imagemUrl;
-        });
-      
-        const deleteButton = document.createElement("button");
-        deleteButton.setAttribute("type", "button");
-        deleteButton.classList.add("btn", "btn-danger", "mx-2");
-        deleteButton.textContent = "Excluir";
-        deleteButton.addEventListener("click", function () {
-          if (confirm("Tem certeza que deseja excluir este animal perdido?")) {
-            deletePet(pet.id, loadAndDisplayPets);
-          }
-        });
-      
-        const editDeleteContainer = document.createElement("div");
-        editDeleteContainer.classList.add("edit-delete-container");
-        editDeleteContainer.appendChild(editButton);
-        editDeleteContainer.appendChild(deleteButton);
-      
         box.appendChild(h3);
         box.appendChild(species);
         box.appendChild(address);
         box.appendChild(description);
         box.appendChild(contacts);
         box.appendChild(status);
-        box.appendChild(editDeleteContainer); // Adicionando o container com os botões
+
+        if (pet.userId === userId) {
+          const editDeleteContainer = document.createElement("div");
+          editDeleteContainer.classList.add("edit-delete-container");
+
+          const editButton = document.createElement("button");
+          editButton.setAttribute("type", "button");
+          editButton.classList.add("btn", "btn-custom-green");
+          editButton.setAttribute("data-bs-toggle", "modal");
+          editButton.setAttribute("data-bs-target", "#editModal");
+          editButton.textContent = "Editar";
+      
+          editButton.addEventListener("click", function () {
+            document.getElementById("editId").value = pet.id;
+            document.getElementById("status").value = pet.status;
+            document.getElementById("especie").value = pet.especie;
+            document.getElementById("genero").value = pet.genero;
+            document.getElementById("nome").value = pet.nome;
+            document.getElementById("endereco").value = pet.endereco;
+            document.getElementById("contatos").value = pet.contatos;
+            document.getElementById("descricao").value = pet.descricao;
+            document.getElementById("imagemUrl").value = pet.imagemUrl;
+          });
+          editDeleteContainer.appendChild(editButton);
+      
+          const deleteButton = document.createElement("button");
+          deleteButton.setAttribute("type", "button");
+          deleteButton.classList.add("btn", "btn-danger", "mx-2");
+          deleteButton.textContent = "Excluir";
+          deleteButton.addEventListener("click", function () {
+            if (confirm("Tem certeza que deseja excluir este animal perdido?")) {
+              deletePet(pet.id, loadAndDisplayPets);
+            }
+          });
+          editDeleteContainer.appendChild(deleteButton);
+
+          box.appendChild(editDeleteContainer); // Adicionando o container com os botões
+        }
       
         content.appendChild(petImage);
         content.appendChild(box);
       
         resultsSection.appendChild(content);
       });
-
     })
     .catch((error) => console.error("Erro ao carregar os dados:", error));
 }
 
 document.getElementById("btnFiltrar").addEventListener("click", () => {
-  const selectedStatus = document.querySelector(
-    'input[name="status"]:checked'
-  )?.value;
-  const selectedTipo = document.querySelector(
-    'input[name="tipo"]:checked'
-  )?.value;
+  const selectedStatus = document.querySelector('input[name="status"]:checked')?.value;
+  const selectedTipo = document.querySelector('input[name="tipo"]:checked')?.value;
   loadAndDisplayPets(selectedStatus, selectedTipo);
 });
 
 document.getElementById("btnResetar").addEventListener("click", () => {
-  document
-    .querySelectorAll('input[name="status"]')
-    .forEach((input) => (input.checked = false));
-  document
-    .querySelectorAll('input[name="tipo"]')
-    .forEach((input) => (input.checked = false));
+  document.querySelectorAll('input[name="status"]').forEach((input) => (input.checked = false));
+  document.querySelectorAll('input[name="tipo"]').forEach((input) => (input.checked = false));
   loadAndDisplayPets();
 });
 
@@ -161,6 +159,22 @@ function updatePet(id, pet) {
     .catch(error => {
       console.error('Erro ao atualizar Anúncio via API JSONServer:', error);
       displayMessage("Erro ao atualizar Anúncio");
+    });
+}
+
+function deletePet(id, refreshFunction) {
+  fetch(`${apiUrl}/${id}`, {
+    method: 'DELETE',
+  })
+    .then(response => response.json())
+    .then(data => {
+      console.log("Anúncio removido com sucesso:", data);
+      displayMessage("Anúncio removido com sucesso");
+      if (refreshFunction) refreshFunction();
+    })
+    .catch(error => {
+      console.error('Erro ao remover Anúncio via API JSONServer:', error);
+      displayMessage("Erro ao remover Anúncio");
     });
 }
